Rename normal ODHO into tappable one
