+++ osu.Game.Rulesets.Osu/Difficulty/Evaluators/FlashlightEvaluator.cs
@@ -34,7 +34,7 @@ namespace osu.Game.Rulesets.Osu.Difficulty.Evaluators
         /// </summary>
         public static double EvaluateDifficultyOf(DifficultyHitObject current, IReadOnlyList<Mod> mods)
         {
-            if (current is not OsuDifficultyHitObject osuCurrent)
+            if (current is not OsuTappableDifficultyHitObject osuCurrent)
                 return 0;
 
             var osuHitObject = (OsuHitObject)(osuCurrent.BaseObject);
@@ -45,14 +45,14 @@ namespace osu.Game.Rulesets.Osu.Difficulty.Evaluators
 
             double result = 0.0;
 
-            OsuDifficultyHitObject lastObj = osuCurrent;
+            OsuTappableDifficultyHitObject lastObj = osuCurrent;
 
             double angleRepeatCount = 0.0;
 
             // This is iterating backwards in time from the current object.
-            for (int i = 0; i < Math.Min(osuCurrent.IndexMain, 10); i++)
+            for (int i = 0; i < Math.Min(osuCurrent.IndexTappable, 10); i++)
             {
-                var currentObj = osuCurrent.PreviousMain(i)!;
+                var currentObj = osuCurrent.PreviousTappable(i)!;
                 var currentHitObject = (OsuHitObject)(currentObj.BaseObject);
 
                 cumulativeStrainTime += lastObj.AdjustedDeltaTime;
