+++ osu.Game.Rulesets.Osu/Difficulty/Preprocessing/OsuTappableDifficultyHitObject.cs
@@ -1,0 +1,397 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using osu.Game.Rulesets.Difficulty.Preprocessing;
+using osu.Game.Rulesets.Objects;
+using osu.Game.Rulesets.Osu.Mods;
+using osu.Game.Rulesets.Osu.Objects;
+using osu.Game.Rulesets.Scoring;
+using osuTK;
+
+namespace osu.Game.Rulesets.Osu.Difficulty.Preprocessing
+{
+    public class OsuTappableDifficultyHitObject : OsuDifficultyHitObject
+    {
+        /// <summary>
+        /// A distance by which all distances should be scaled in order to assume a uniform circle size.
+        /// </summary>
+        public const int NORMALISED_RADIUS = 50; // Change radius to 50 to make 100 the diameter. Easier for mental maths.
+
+        public const int NORMALISED_DIAMETER = NORMALISED_RADIUS * 2;
+
+        public const int MIN_DELTA_TIME = 25;
+
+        private const float maximum_slider_radius = NORMALISED_RADIUS * 2.4f;
+        private const float assumed_slider_radius = NORMALISED_RADIUS * 1.8f;
+
+        private readonly IReadOnlyList<OsuTappableDifficultyHitObject> tappableDifficultyHitObjects;
+        private readonly OsuHitObject tappableMainObject;
+
+        protected new OsuHitObject BaseObject => (OsuHitObject)base.BaseObject;
+        protected new OsuHitObject LastObject => (OsuHitObject)base.LastObject;
+
+        /// <summary>
+        /// The index of this <see cref="OsuTappableDifficultyHitObject"/> in the list of tappable <see cref="OsuTappableDifficultyHitObject"/>s.
+        /// </summary>
+        public int IndexTappable;
+
+        /// <summary>
+        /// <see cref="DifficultyHitObject.DeltaTime"/> between tappable objects.
+        /// </summary>
+        public readonly double DeltaTimeTappable;
+
+        /// <summary>
+        /// <see cref="DifficultyHitObject.DeltaTime"/> between tappable objects capped to a minimum of <see cref="MIN_DELTA_TIME"/>ms.
+        /// </summary>
+        public readonly double AdjustedDeltaTime;
+
+        /// <summary>
+        /// Normalised distance from the "lazy" end position of the previous <see cref="OsuTappableDifficultyHitObject"/> to the start position of this <see cref="OsuTappableDifficultyHitObject"/>.
+        /// <para>
+        /// The "lazy" end position is the position at which the cursor ends up if the previous hitobject is followed with as minimal movement as possible (i.e. on the edge of slider follow circles).
+        /// </para>
+        /// </summary>
+        public double LazyJumpDistance { get; private set; }
+
+        /// <summary>
+        /// Normalised shortest distance to consider for a jump between the previous <see cref="OsuTappableDifficultyHitObject"/> and this <see cref="OsuTappableDifficultyHitObject"/>.
+        /// </summary>
+        /// <remarks>
+        /// This is bounded from above by <see cref="LazyJumpDistance"/>, and is smaller than the former if a more natural path is able to be taken through the previous <see cref="OsuTappableDifficultyHitObject"/>.
+        /// </remarks>
+        /// <example>
+        /// Suppose a linear slider - circle pattern.
+        /// <br />
+        /// Following the slider lazily (see: <see cref="LazyJumpDistance"/>) will result in underestimating the true end position of the slider as being closer towards the start position.
+        /// As a result, <see cref="LazyJumpDistance"/> overestimates the jump distance because the player is able to take a more natural path by following through the slider to its end,
+        /// such that the jump is felt as only starting from the slider's true end position.
+        /// <br />
+        /// Now consider a slider - circle pattern where the circle is stacked along the path inside the slider.
+        /// In this case, the lazy end position correctly estimates the true end position of the slider and provides the more natural movement path.
+        /// </example>
+        public double MinimumJumpDistance { get; private set; }
+
+        /// <summary>
+        /// The time taken to travel through <see cref="MinimumJumpDistance"/>, with a minimum value of 25ms.
+        /// </summary>
+        public double MinimumJumpTime { get; private set; }
+
+        /// <summary>
+        /// Normalised distance between the start and end position of this <see cref="OsuTappableDifficultyHitObject"/>.
+        /// </summary>
+        public double TravelDistance { get; private set; }
+
+        /// <summary>
+        /// The time taken to travel through <see cref="TravelDistance"/>, with a minimum value of 25ms for <see cref="Slider"/> objects.
+        /// </summary>
+        public double TravelTime { get; private set; }
+
+        /// <summary>
+        /// The position of the cursor at the point of completion of this <see cref="OsuTappableDifficultyHitObject"/> if it is a <see cref="Slider"/>
+        /// and was hit with as few movements as possible.
+        /// </summary>
+        public Vector2? LazyEndPosition { get; private set; }
+
+        /// <summary>
+        /// The distance travelled by the cursor upon completion of this <see cref="OsuTappableDifficultyHitObject"/> if it is a <see cref="Slider"/>
+        /// and was hit with as few movements as possible.
+        /// </summary>
+        public double LazyTravelDistance { get; private set; }
+
+        /// <summary>
+        /// The time taken by the cursor upon completion of this <see cref="OsuTappableDifficultyHitObject"/> if it is a <see cref="Slider"/>
+        /// and was hit with as few movements as possible.
+        /// </summary>
+        public double LazyTravelTime { get; private set; }
+
+        /// <summary>
+        /// Angle the player has to take to hit this <see cref="OsuTappableDifficultyHitObject"/>.
+        /// Calculated as the angle between the circles (current-2, current-1, current).
+        /// </summary>
+        public double? Angle { get; private set; }
+
+        /// <summary>
+        /// Retrieves the full hit window for a Great <see cref="HitResult"/>.
+        /// </summary>
+        public double HitWindowGreat { get; private set; }
+
+        /// <summary>
+        /// Selective bonus for maps with higher circle size.
+        /// </summary>
+        public double SmallCircleBonus { get; private set; }
+
+        private readonly OsuTappableDifficultyHitObject? lastLastDifficultyObject;
+        private readonly OsuTappableDifficultyHitObject? lastDifficultyObject;
+
+        public OsuTappableDifficultyHitObject(HitObject hitObject, HitObject lastObject, HitObject lastMainObject, double clockRate, List<DifficultyHitObject> objects, int index, List<OsuTappableDifficultyHitObject> tappableObjects, int tappableIndex)
+            : base(hitObject, lastObject, clockRate, objects, index)
+        {
+            this.tappableMainObject = (OsuHitObject)lastMainObject;
+            tappableDifficultyHitObjects = tappableObjects;
+            IndexTappable = tappableIndex;
+
+            lastLastDifficultyObject = index > 1 ? tappableObjects[tappableIndex - 2] : null;
+            lastDifficultyObject = index > 0 ? tappableObjects[tappableIndex - 1] : null;
+
+            DeltaTimeTappable = (hitObject.StartTime - lastMainObject.StartTime) / clockRate;
+
+            // Capped to 25ms to prevent difficulty calculation breaking from simultaneous objects.
+            AdjustedDeltaTime = Math.Max(DeltaTimeTappable, MIN_DELTA_TIME);
+
+            SmallCircleBonus = Math.Max(1.0, 1.0 + (30 - BaseObject.Radius) / 40);
+
+            if (BaseObject is Slider sliderObject)
+            {
+                HitWindowGreat = 2 * sliderObject.HeadCircle.HitWindows.WindowFor(HitResult.Great) / clockRate;
+            }
+            else
+            {
+                HitWindowGreat = 2 * BaseObject.HitWindows.WindowFor(HitResult.Great) / clockRate;
+            }
+
+            computeSliderCursorPosition();
+            setDistances(clockRate);
+        }
+
+        public OsuTappableDifficultyHitObject? PreviousTappable(int backwardsIndex)
+        {
+            int index = IndexTappable - (backwardsIndex + 1);
+            return index >= 0 && index < tappableDifficultyHitObjects.Count ? tappableDifficultyHitObjects[index] : default;
+        }
+
+        public OsuTappableDifficultyHitObject? NextTappable(int forwardsIndex)
+        {
+            int index = IndexTappable + (forwardsIndex + 1);
+            return index >= 0 && index < tappableDifficultyHitObjects.Count ? tappableDifficultyHitObjects[index] : default;
+        }
+
+        public double OpacityAt(double time, bool hidden)
+        {
+            if (time > BaseObject.StartTime)
+            {
+                // Consider a hitobject as being invisible when its start time is passed.
+                // In reality the hitobject will be visible beyond its start time up until its hittable window has passed,
+                // but this is an approximation and such a case is unlikely to be hit where this function is used.
+                return 0.0;
+            }
+
+            double fadeInStartTime = BaseObject.StartTime - BaseObject.TimePreempt;
+            double fadeInDuration = BaseObject.TimeFadeIn;
+
+            if (hidden)
+            {
+                // Taken from OsuModHidden.
+                double fadeOutStartTime = BaseObject.StartTime - BaseObject.TimePreempt + BaseObject.TimeFadeIn;
+                double fadeOutDuration = BaseObject.TimePreempt * OsuModHidden.FADE_OUT_DURATION_MULTIPLIER;
+
+                return Math.Min
+                (
+                    Math.Clamp((time - fadeInStartTime) / fadeInDuration, 0.0, 1.0),
+                    1.0 - Math.Clamp((time - fadeOutStartTime) / fadeOutDuration, 0.0, 1.0)
+                );
+            }
+
+            return Math.Clamp((time - fadeInStartTime) / fadeInDuration, 0.0, 1.0);
+        }
+
+        /// <summary>
+        /// Returns how possible is it to doubletap this object together with the next one and get perfect judgement in range from 0 to 1
+        /// </summary>
+        public double GetDoubletapness(OsuTappableDifficultyHitObject? osuNextObj)
+        {
+            if (osuNextObj != null)
+            {
+                double currDeltaTime = Math.Max(1, DeltaTime);
+                double nextDeltaTime = Math.Max(1, osuNextObj.DeltaTime);
+
+                double deltaDifference = Math.Abs(nextDeltaTime - currDeltaTime);
+
+                double speedRatio = currDeltaTime / Math.Max(currDeltaTime, deltaDifference);
+                double windowRatio = Math.Pow(Math.Min(1, currDeltaTime / HitWindowGreat), 5);
+
+                return 1.0 - Math.Pow(speedRatio, 1 - windowRatio);
+            }
+
+            return 0;
+        }
+
+        private void setDistances(double clockRate)
+        {
+            if (BaseObject is Slider currentSlider)
+            {
+                // Bonus for repeat sliders until a better per nested object strain system can be achieved.
+                TravelDistance = LazyTravelDistance * Math.Pow(1 + currentSlider.RepeatCount / 2.5, 1.0 / 2.5);
+                TravelTime = Math.Max(LazyTravelTime / clockRate, MIN_DELTA_TIME);
+            }
+
+            // We will scale distances by this factor, so we can assume a uniform CircleSize among beatmaps.
+            float scalingFactor = NORMALISED_RADIUS / (float)BaseObject.Radius;
+
+            Vector2 lastCursorPosition = lastDifficultyObject != null ? getEndCursorPosition(lastDifficultyObject) : tappableMainObject.StackedPosition;
+
+            LazyJumpDistance = (BaseObject.StackedPosition * scalingFactor - lastCursorPosition * scalingFactor).Length;
+            MinimumJumpTime = AdjustedDeltaTime;
+            MinimumJumpDistance = LazyJumpDistance;
+
+            if (tappableMainObject is Slider lastSlider && lastDifficultyObject != null)
+            {
+                double lastTravelTime = Math.Max(lastDifficultyObject.LazyTravelTime / clockRate, MIN_DELTA_TIME);
+                MinimumJumpTime = Math.Max(AdjustedDeltaTime - lastTravelTime, MIN_DELTA_TIME);
+
+                //
+                // There are two types of slider-to-object patterns to consider in order to better approximate the real movement a player will take to jump between the hitobjects.
+                //
+                // 1. The anti-flow pattern, where players cut the slider short in order to move to the next hitobject.
+                //
+                //      <======o==>  ← slider
+                //             |     ← most natural jump path
+                //             o     ← a follow-up hitcircle
+                //
+                // In this case the most natural jump path is approximated by LazyJumpDistance.
+                //
+                // 2. The flow pattern, where players follow through the slider to its visual extent into the next hitobject.
+                //
+                //      <======o==>---o
+                //                  ↑
+                //        most natural jump path
+                //
+                // In this case the most natural jump path is better approximated by a new distance called "tailJumpDistance" - the distance between the slider's tail and the next hitobject.
+                //
+                // Thus, the player is assumed to jump the minimum of these two distances in all cases.
+                //
+
+                float tailJumpDistance = Vector2.Subtract(lastSlider.TailCircle.StackedPosition, BaseObject.StackedPosition).Length * scalingFactor;
+                MinimumJumpDistance = Math.Max(0, Math.Min(LazyJumpDistance - (maximum_slider_radius - assumed_slider_radius), tailJumpDistance - maximum_slider_radius));
+            }
+
+            if (lastLastDifficultyObject != null && lastLastDifficultyObject.BaseObject is not Spinner)
+            {
+                Vector2 lastLastCursorPosition = getEndCursorPosition(lastLastDifficultyObject);
+
+                Vector2 v1 = lastLastCursorPosition - tappableMainObject.StackedPosition;
+                Vector2 v2 = BaseObject.StackedPosition - lastCursorPosition;
+
+                float dot = Vector2.Dot(v1, v2);
+                float det = v1.X * v2.Y - v1.Y * v2.X;
+
+                Angle = Math.Abs(Math.Atan2(det, dot));
+            }
+        }
+
+        private void computeSliderCursorPosition()
+        {
+            if (BaseObject is not Slider slider)
+                return;
+
+            if (LazyEndPosition != null)
+                return;
+
+            // TODO: This commented version is actually correct by the new lazer implementation, but intentionally held back from
+            // difficulty calculator to preserve known behaviour.
+            // double trackingEndTime = Math.Max(
+            //     // SliderTailCircle always occurs at the final end time of the slider, but the player only needs to hold until within a lenience before it.
+            //     slider.Duration + SliderEventGenerator.TAIL_LENIENCY,
+            //     // There's an edge case where one or more ticks/repeats fall within that leniency range.
+            //     // In such a case, the player needs to track until the final tick or repeat.
+            //     slider.NestedHitObjects.LastOrDefault(n => n is not SliderTailCircle)?.StartTime ?? double.MinValue
+            // );
+
+            double trackingEndTime = Math.Max(
+                slider.StartTime + slider.Duration + SliderEventGenerator.TAIL_LENIENCY,
+                slider.StartTime + slider.Duration / 2
+            );
+
+            IList<HitObject> nestedObjects = slider.NestedHitObjects;
+
+            SliderTick? lastRealTick = null;
+
+            foreach (var hitobject in slider.NestedHitObjects)
+            {
+                if (hitobject is SliderTick tick)
+                    lastRealTick = tick;
+            }
+
+            if (lastRealTick?.StartTime > trackingEndTime)
+            {
+                trackingEndTime = lastRealTick.StartTime;
+
+                // When the last tick falls after the tracking end time, we need to re-sort the nested objects
+                // based on time. This creates a somewhat weird ordering which is counter to how a user would
+                // understand the slider, but allows a zero-diff with known diffcalc output.
+                //
+                // To reiterate, this is definitely not correct from a difficulty calculation perspective
+                // and should be revisited at a later date (likely by replacing this whole code with the commented
+                // version above).
+                List<HitObject> reordered = nestedObjects.ToList();
+
+                reordered.Remove(lastRealTick);
+                reordered.Add(lastRealTick);
+
+                nestedObjects = reordered;
+            }
+
+            LazyTravelTime = trackingEndTime - slider.StartTime;
+
+            double endTimeMin = LazyTravelTime / slider.SpanDuration;
+            if (endTimeMin % 2 >= 1)
+                endTimeMin = 1 - endTimeMin % 1;
+            else
+                endTimeMin %= 1;
+
+            LazyEndPosition = slider.StackedPosition + slider.Path.PositionAt(endTimeMin); // temporary lazy end position until a real result can be derived.
+
+            Vector2 currCursorPosition = slider.StackedPosition;
+
+            double scalingFactor = NORMALISED_RADIUS / slider.Radius; // lazySliderDistance is coded to be sensitive to scaling, this makes the maths easier with the thresholds being used.
+
+            for (int i = 1; i < nestedObjects.Count; i++)
+            {
+                var currMovementObj = (OsuHitObject)nestedObjects[i];
+
+                Vector2 currMovement = Vector2.Subtract(currMovementObj.StackedPosition, currCursorPosition);
+                double currMovementLength = scalingFactor * currMovement.Length;
+
+                // Amount of movement required so that the cursor position needs to be updated.
+                double requiredMovement = assumed_slider_radius;
+
+                if (i == nestedObjects.Count - 1)
+                {
+                    // The end of a slider has special aim rules due to the relaxed time constraint on position.
+                    // There is both a lazy end position as well as the actual end slider position. We assume the player takes the simpler movement.
+                    // For sliders that are circular, the lazy end position may actually be farther away than the sliders true end.
+                    // This code is designed to prevent buffing situations where lazy end is actually a less efficient movement.
+                    Vector2 lazyMovement = Vector2.Subtract((Vector2)LazyEndPosition, currCursorPosition);
+
+                    if (lazyMovement.Length < currMovement.Length)
+                        currMovement = lazyMovement;
+
+                    currMovementLength = scalingFactor * currMovement.Length;
+                }
+                else if (currMovementObj is SliderRepeat)
+                {
+                    // For a slider repeat, assume a tighter movement threshold to better assess repeat sliders.
+                    requiredMovement = NORMALISED_RADIUS;
+                }
+
+                if (currMovementLength > requiredMovement)
+                {
+                    // this finds the positional delta from the required radius and the current position, and updates the currCursorPosition accordingly, as well as rewarding distance.
+                    currCursorPosition = Vector2.Add(currCursorPosition, Vector2.Multiply(currMovement, (float)((currMovementLength - requiredMovement) / currMovementLength)));
+                    currMovementLength *= (currMovementLength - requiredMovement) / currMovementLength;
+                    LazyTravelDistance += currMovementLength;
+                }
+
+                if (i == nestedObjects.Count - 1)
+                    LazyEndPosition = currCursorPosition;
+            }
+        }
+
+        private Vector2 getEndCursorPosition(OsuTappableDifficultyHitObject difficultyHitObject)
+        {
+            return difficultyHitObject.LazyEndPosition ?? difficultyHitObject.BaseObject.StackedPosition;
+        }
+    }
+}
