+++ osu.Game.Rulesets.Osu/Difficulty/Skills/Speed.cs
@@ -40,13 +40,13 @@ namespace osu.Game.Rulesets.Osu.Difficulty.Skills
         protected override double StrainValueAt(DifficultyHitObject current)
         {
             // If object is not tappable - just skip
-            if (current is not OsuDifficultyHitObject)
+            if (current is not OsuTappableDifficultyHitObject)
             {
                 currentStrain *= strainDecay(current.DeltaTime);
                 return currentStrain * currentRhythm;
             }
 
-            currentStrain *= strainDecay(Math.Max(current.DeltaTime, OsuDifficultyHitObject.MIN_DELTA_TIME));
+            currentStrain *= strainDecay(Math.Max(current.DeltaTime, OsuTappableDifficultyHitObject.MIN_DELTA_TIME));
             currentStrain += SpeedEvaluator.EvaluateDifficultyOf(current, Mods) * skillMultiplier;
 
             currentRhythm = RhythmEvaluator.EvaluateDifficultyOf(current);
